pyproject.toml
[project]
name = "automacao-fipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
]
